+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## 0.4.0
 ### Added
 - Keep browsers open while scanning in automation mode
+- Help to explain how to set the options from the command line.
+
+### Changed
+- Options to save the rule settings by ID.
 
 ## 0.3.0
 ### Added